+++ src/cli/commands/public-cmds/init-cmd.ts
@@ -12,7 +12,8 @@ import shouldShowInteractive from '../../../interactive/utils/should-show-intera
 
 export default class Init extends Command {
   name = 'init [path]';
-  description = `initialize an empty bit scope\n  https://${BASE_DOCS_DOMAIN}/docs/apis/cli-all#init`;
+  skipWorkspace = true;
+  description = `initialize an empty bit scope\n  https://${BASE_DOCS_DOMAIN}/docs/workspace`;
   alias = '';
   // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
   opts = [