+++ src/cli/commands/public-cmds/show-cmd.ts
@@ -8,7 +8,7 @@ import { DependenciesInfo } from '../../../scope/graph/scope-graph';
 
 export default class Show extends Command {
   name = 'show <id>';
-  description = `show component overview.\n https://${BASE_DOCS_DOMAIN}/docs/apis/cli-all#show`;
+  description = `show component overview.\n https://${BASE_DOCS_DOMAIN}/docs/view#show`;
   alias = '';
   // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
   opts = [
@@ -23,6 +23,7 @@ export default class Show extends Command {
   ];
   loader = true;
   migration = true;
+  skipWorkspace = true;
 
   action(
     [id]: [string],