+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [14.4.4-dev.2] - 2019-11-10
+
 - fix overrides to respect dependencies received by a compiler/tester
 - improve loader when building components
 - [#2101](https://github.com/teambit/bit/issues/2101) suggest matching commands